backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.16
uvicorn==0.24.0
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0